documentation added

+++ str/match_string.h
@@ -4,6 +4,17 @@
 #include "kernel_definitions.h"
 #include "strcmp.h"
 
-ssize_t match_string(const char * const *array, size_t n, const char *string);
+
+/**
+ * match_string - matches given string in an array
+ * @array:	array of strings
+ * @n:		number of strings in the array or -1 for NULL terminated arrays
+ * @string:	string to match with
+ *
+ * Return:
+ * index of a @string in the @array if matches, or %-EINVAL otherwise.
+ */
+
+int match_string(const char * const *array, size_t n, const char *string);
 
 #endif // __MATCH_STRING_H__
